==> 00_get_data.ipynb <==
## Data Collection (00)

**Import Libraries**

pwd

import pandas as pd
import requests
import time
import datetime
import warnings

import sys

if not sys.warnoptions:
    import warnings
    warnings.simplefilter("ignore")

**Get Data**

def get_comments(subreddit, n_iter, epoch_right_now): # subreddit name and number of times function should run
    
    # store base url variable
    base_url = 'https://api.pushshift.io/reddit/search/submission/?subreddit='
    # instantiate empty list    
    df_list = []
    # save current epoch, used to iterate in reverse through time
    current_time = epoch_right_now
    
    # set up for loop
    for post in range(n_iter):
        
        
        # instantiate get request
        res = requests.get(
            
            # requests.get takes base_url and params
            base_url,
            # parameters for get request
            params={
                
                # specify subreddit
                "subreddit": subreddit,
                # specify number of posts to pull
                "size": 1000,   
                # ???
                "lang": True,
                # pull everything from current time backward
                "before": current_time
            }
        )
        
        # take data from most recent request, store as df
        df = pd.DataFrame(res.json()['data'])
        # pull specific columns from dataframe for analysis
        df = df.loc[:, ['title',
                        'created_utc', 
                        'selftext',
                        'subreddit',
                        'id',
                        'media_only']]
        # apend to empty dataframe list
        df_list.append(df)
        # set current time counter back to last epoch in recently grabbed df
        current_time = df['created_utc'].min()

    # return one dataframe for all requests
    return pd.concat(df_list, axis=0)

# Adapated from Tim Book's Lesson Example

**Write to DataFrame**

# Cooking
df_cooking = get_comments('Cooking',100,1587593906)

# Baking
df_baking = get_comments('Baking',100,1587593906)

**Save to CSV**

# Write to DataFrame
df_cooking.to_csv('cooking.csv', index=False)

# Write to DataFrame
df_baking.to_csv('baking.csv', index=False)

**Notes**

- Explore how the request.get(url, params) works: specifically url + subreddit syntax
- Research methods of exporting DataFrames with the same name with an appended statement to indicate time generated
