==> 01_clean_data.ipynb <==
## Data Exploration and Cleaning (01)

**Import Libraries**

import pandas as pd
import numpy as np
from pprint import pprint
from sara import eda, eda_unique

**Read in Data**

df_cooking = pd.read_csv('data/cooking.csv')
df_baking = pd.read_csv('data/baking.csv')

**Create Single DataFrame**

# Concat df_cooking and df_baking
df_reddit = pd.concat([df_cooking,df_baking])
# Check shape
#pprint(df_reddit.shape)
# Check subreddits by class
#pprint(df_reddit['subreddit'].value_counts(normalize=True))
#eda(df_reddit)

**Remove Nulls/NaNs**

# Clean title
df_reddit['title'].isna().sum()
df_reddit.dropna(subset=['title'],inplace=True)

# Clean selftext
df_reddit['selftext'].isna().sum()
df_reddit.dropna(subset=['selftext'],inplace=True)

**Read Cleaned Data into CSV**

df_reddit.to_csv('./data/reddit_cleaned.csv',index=False)

**Notes**

- Most cleaning occurs in the processing stage as part of NLP
